+++ src/program/ui/RamWatchEditWindow.cpp
@@ -22,6 +22,7 @@
 #include "ramsearch/TypeIndex.h"
 #include "ramsearch/IRamWatchDetailed.h"
 #include "ramsearch/RamWatchDetailed.h"
+#include "ramsearch/RamWatchDetailedBuilder.h"
 
 #include <QtWidgets/QDialogButtonBox>
 #include <QtWidgets/QPushButton>
@@ -296,38 +297,7 @@ void RamWatchEditWindow::slotSave()
         reject();
 
     /* Build the ram watch using the right type as template */
-    switch (typeBox->currentIndex()) {
-        case RamUnsignedChar:
-            ramwatch.reset(new RamWatchDetailed<unsigned char>(addr));
-            break;
-        case RamChar:
-            ramwatch.reset(new RamWatchDetailed<char>(addr));
-            break;
-        case RamUnsignedShort:
-            ramwatch.reset(new RamWatchDetailed<unsigned short>(addr));
-            break;
-        case RamShort:
-            ramwatch.reset(new RamWatchDetailed<short>(addr));
-            break;
-        case RamUnsignedInt:
-            ramwatch.reset(new RamWatchDetailed<unsigned int>(addr));
-            break;
-        case RamInt:
-            ramwatch.reset(new RamWatchDetailed<int>(addr));
-            break;
-        case RamUnsignedLong:
-            ramwatch.reset(new RamWatchDetailed<uint64_t>(addr));
-            break;
-        case RamLong:
-            ramwatch.reset(new RamWatchDetailed<int64_t>(addr));
-            break;
-        case RamFloat:
-            ramwatch.reset(new RamWatchDetailed<float>(addr));
-            break;
-        case RamDouble:
-            ramwatch.reset(new RamWatchDetailed<double>(addr));
-            break;
-    }
+    ramwatch.reset(RamWatchDetailedBuilder::new_watch(addr, typeBox->currentIndex()));
 
     ramwatch->hex = (displayBox->currentIndex() == 1);
     ramwatch->label = labelInput->text().toStdString();