+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 * Add Shift modifier for text input
 * Handle softlock inside snd_pcm_wait()
 * Set corresponding radio button when changing a screen resolution value
+* Save and intersect pointermaps
 
 ### Changed